OAuth, restructure init

+++ __SYSDefault/InitDocs/InitOperations/ArrivalAns.js
@@ -15,7 +15,8 @@ module.exports = async () => {
   
   let db = await _remote.BaseDB();
   let res;
-  
+  let dbName;
+
   try{
     dbName = ConfigDocs.DBNAME.ArrivalAns;
     res = await db.DestroyDatabase(dbName);
@@ -24,31 +25,63 @@ module.exports = async () => {
       throw new Error(res.payload.Error)
     }
 
-    await Promise.all(_.map(DBDocs.ArrivalAns, async (o, i) => {
+    await Promise.all(_.map(DBDocs.ArrivalAns.Template, async (o, i) => {
       res = await db.Insert(dbName, o);
     }));
 
-    dbName = ConfigDocs.DBNAME.Footnote;
+    ///
+
+    dbName = ConfigDocs.DBNAME.APProc;
     res = await db.DestroyDatabase(dbName);
     res = await db.CreateDatabase(dbName);
     if(!res.Success){
       throw new Error(res.payload.Error)
     }
 
-    await Promise.all(_.map(DBDocs.Footnote, async (o, i) => {
+    await Promise.all(_.map(DBDocs.ArrivalAns.Components.APProc, async (o, i) => {
       res = await db.Insert(dbName, o);
     }));
 
-    dbName = ConfigDocs.DBNAME.DefaultQ;
+    ///
+
+    dbName = ConfigDocs.DBNAME.DOCReq;
     res = await db.DestroyDatabase(dbName);
     res = await db.CreateDatabase(dbName);
     if(!res.Success){
       throw new Error(res.payload.Error)
     }
 
-    await Promise.all(_.map(DBDocs.DefaultQ, async (o, i) => {
+    await Promise.all(_.map(DBDocs.ArrivalAns.Components.DOCReq, async (o, i) => {
       res = await db.Insert(dbName, o);
     }));
+
+    ///
+
+    dbName = ConfigDocs.DBNAME.ENTReq;
+    res = await db.DestroyDatabase(dbName);
+    res = await db.CreateDatabase(dbName);
+    if(!res.Success){
+      throw new Error(res.payload.Error)
+    }
+
+    await Promise.all(_.map(DBDocs.ArrivalAns.Components.ENTReq, async (o, i) => {
+      res = await db.Insert(dbName, o);
+    }));
+
+    ///
+
+    dbName = ConfigDocs.DBNAME.QUAReq;
+    res = await db.DestroyDatabase(dbName);
+    res = await db.CreateDatabase(dbName);
+    if(!res.Success){
+      throw new Error(res.payload.Error)
+    }
+
+    await Promise.all(_.map(DBDocs.ArrivalAns.Components.QUAReq, async (o, i) => {
+      res = await db.Insert(dbName, o);
+    }));
+
+    ///
 
     dbName = ConfigDocs.DBNAME.Tips;
     res = await db.DestroyDatabase(dbName);
@@ -57,9 +90,24 @@ module.exports = async () => {
       throw new Error(res.payload.Error)
     }
 
-    await Promise.all(_.map(DBDocs.Tips, async (o, i) => {
+    await Promise.all(_.map(DBDocs.ArrivalAns.Components.Tips, async (o, i) => {
       res = await db.Insert(dbName, o);
     }));
+
+    ///
+
+    dbName = ConfigDocs.DBNAME.Footnote;
+    res = await db.DestroyDatabase(dbName);
+    res = await db.CreateDatabase(dbName);
+    if(!res.Success){
+      throw new Error(res.payload.Error)
+    }
+
+    await Promise.all(_.map(DBDocs.ArrivalAns.Miscellaneous.Footnote, async (o, i) => {
+      res = await db.Insert(dbName, o);
+    }));
+
+    ///
 
     dbName = ConfigDocs.DBNAME.IconDocs;
     res = await db.DestroyDatabase(dbName);
@@ -68,7 +116,7 @@ module.exports = async () => {
       throw new Error(res.payload.Error)
     }
 
-    await Promise.all(_.map(DBDocs.IconDocs, async (o, i) => {
+    await Promise.all(_.map(DBDocs.ArrivalAns.Miscellaneous.IconDocs, async (o, i) => {
       res = await db.Insert(dbName, o);
     }));
